+++ src/Compatibility/MenuFrameworkRuntime.h
@@ -10,6 +10,7 @@ namespace whereabouts
     struct MenuFrameworkRuntimeProbe
     {
         bool moduleLoaded{false};
+        MenuFrameworkProvider provider{MenuFrameworkProvider::Unknown};
         std::optional<MenuFrameworkVersion> fixedVersion;
         bool requiredExportsAvailable{false};
         std::string missingRequiredExport;